Mention Swift container setup
The hastexo XBlock requires Swift storage. With Ansible 2.x, this
would be easy to automate by use of the os_object module, but Open edX
is stuck on their own fork of Ansible 1.9.3, so that is not an
option. Hence, we need to instruct users to create the required
container manually.